+++ tests/test_visitors/conftest.py
@@ -11,15 +11,6 @@ from wemake_python_styleguide.compat import maybe_set_parent
 from wemake_python_styleguide.options import defaults
 from wemake_python_styleguide.visitors.base.visitor import BaseNodeVisitor
 
-Options = namedtuple('options', [
-    'max_arguments',
-    'max_expressions',
-    'max_local_variables',
-    'max_returns',
-    'min_variable_length',
-    'max_offset_blocks',
-])
-
 
 @pytest.fixture(scope='session')
 def parse_ast_tree():
@@ -46,18 +37,24 @@ def assert_errors():
 @pytest.fixture(scope='session')
 def options():
     """Returns the options builder."""
-    def factory(**kwargs):
-        default_values = {
-            'max_arguments': defaults.MAX_ARGUMENTS,
-            'max_expressions': defaults.MAX_EXPRESSIONS,
-            'max_local_variables': defaults.MAX_LOCAL_VARIABLES,
-            'max_returns': defaults.MAX_RETURNS,
-            'min_variable_length': defaults.MIN_VARIABLE_LENGTH,
-            'max_offset_blocks': defaults.MAX_OFFSET_BLOCKS,
-        }
+    default_values = {
+        'max_arguments': defaults.MAX_ARGUMENTS,
+        'max_expressions': defaults.MAX_EXPRESSIONS,
+        'max_local_variables': defaults.MAX_LOCAL_VARIABLES,
+        'max_returns': defaults.MAX_RETURNS,
+        'min_variable_length': defaults.MIN_VARIABLE_LENGTH,
+        'max_offset_blocks': defaults.MAX_OFFSET_BLOCKS,
+        'max_elifs': defaults.MAX_ELIFS,
+        'max_module_members': defaults.MAX_MODULE_MEMBERS,
+        'max_methods': defaults.MAX_METHODS,
+    }
 
-        default_values.update(kwargs)
-        return Options(**default_values)
+    Options = namedtuple('options', default_values.keys())
+
+    def factory(**kwargs):
+        final_options = default_values.copy()
+        final_options.update(kwargs)
+        return Options(**final_options)
 
     return factory