+++ wemake_python_styleguide/constants.py
@@ -10,6 +10,8 @@ It also contains some exceptions that we allow to use in our codebase.
 import sys
 from typing import Tuple
 
+# TODO: 'stdin' filename should be a constant
+
 #: List of functions we forbid to use.
 BAD_FUNCTIONS = frozenset((
     # Code generation:
@@ -69,6 +71,9 @@ _BAD_VARIABLE_NAMES: Tuple[str, ...] = (
     'handler',
     'file',
     'klass',
+    'foo',
+    'bar',
+    'baz',
 )
 
 if sys.version_info < (3, 7):  # pragma: no cover
@@ -83,9 +88,10 @@ BAD_VARIABLE_NAMES = frozenset(_BAD_VARIABLE_NAMES)
 
 #: List of magic methods that are forbiden to use.
 BAD_MAGIC_METHODS = frozenset((
-    '__del__',  # since we don't use `del`
-    '__delitem__',  # since we don't use `del`
-    '__delete__',  # since we don't use `del`
+    # Since we don't use `del`:
+    '__del__',
+    '__delitem__',
+    '__delete__',
 
     '__dir__',  # since we don't use `dir()`
     '__delattr__',  # since we don't use `delattr()`